fix(sql): cast PERCENT_RANK a numeric en recalcular_contratos_proveedor
PERCENT_RANK() devuelve double precision; Postgres no tiene overload de
ROUND(double precision, integer), solo ROUND(numeric, integer). Sin el
cast explicito, ingreso_percentil contaminaba el resto de la formula a
double precision y el ROUND fallaba en tiempo de ejecucion.

+++ supabase/migrations/0101_contratos_proveedor_simulados.sql
@@ -70,7 +70,7 @@ BEGIN
   con_percentil AS (
     SELECT
       pm.*,
-      PERCENT_RANK() OVER (ORDER BY pm.total_ingreso) AS ingreso_percentil,
+      (PERCENT_RANK() OVER (ORDER BY pm.total_ingreso))::numeric AS ingreso_percentil,
       CASE
         WHEN (SELECT max_m FROM bounds) > (SELECT min_m FROM bounds)
           THEN 8 - 6 * (pm.margen_pct_bruto - (SELECT min_m FROM bounds)) / ((SELECT max_m FROM bounds) - (SELECT min_m FROM bounds))
